refactor: 중첩된 eventListener 함수로 분리
