+++ src/reset.js
@@ -1,0 +1,9 @@
+import clearUIElements from './View/clear/clearUIElements.js';
+import state from './state.js';
+
+function reset() {
+  state.lottos = [];
+  clearUIElements();
+}
+
+export default reset;
